requirements.txt
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
anthropic==0.11.14
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
pytest==7.4.4
pytest-cov==4.1.0
python-multipart==0.0.6
redis==5.0.1
pandas==2.1.4
numpy==1.26.3
uvloop==0.19.0
httpx==0.26.0
python-dateutil==2.8.2
black==23.12.1
isort==5.13.2
flake8==7.0.0
requests==2.31.0
